tyre_014_back

See merge request ivmm33/tyre_service!19

+++ TyreServiceBack/src/main/java/ru/tyreservice/aggregator/entities/Review.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Embeddable
 public class Review {
     @Column(nullable = false)
-    private Byte ball;
+    private Integer ball;
     @Column(nullable = false)
     private String message;
     @Column(nullable = false)
